+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 * Added `childReactiveState` variant which takes an arbitrary event handler, so the parent ReactiveState doesn't have to implement the whole events interface.
 * Changed `MutableState.beforeUpdate`, `MutableState.afterUpdate` and `MutableState.withSetter` and `State.toMutable` to not be `@Composable`.
 * Removed direct dependency on JUnit 4, so you can choose more freely which JUnit version to use.
+* Added `ifTake` and `unlessTake` inversions of `takeIf` and `takeUnless`.
+* Added `runIf` and `applyIf`.
 
 ## 5.13.0